# move expensive layout sanity check to debug assertions