@lqd 2025-05-15
Although maybe it doesn't really _fix_ #140944 to actually close it, but just removes the worst offender.