@RalfJung 2025-05-15
Oh wait that was already done, I just didn't get the notifications. Odd. I got a notification for the PR description but not all of the other comments...

I don't know a way to cancel the try build or perf run, sorry :shrug: