@lqd 2025-05-17
The "regression" on `nalgebra` opt seems present in other perf runs and is likely noise, like some of the "improvements" in the previous run as well. This PR does improve `lemmy` as seen in the measurements, so we should be good to go.

@bors r=compiler-errors